% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_list_string.R
\name{activityList_format}
\alias{activityList_format}
\title{Output single string of variants from  detected SNVs and CNVs}
\usage{
activityList_format(snv_prep_filepath, cnv_prep_filepath)
}
\arguments{
\item{snv_prep}{}
}
\description{
Output single string of variants from  detected SNVs and CNVs
}
\examples{
ALK: p.D1529E (56.1\%); EGFR: c.1498+22A>T (47.8\%);
}
