% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMissingCols.R
\name{addMissingCols}
\alias{addMissingCols}
\title{Add missing columns to make inconsistent IR output work}
\usage{
addMissingCols(ir_output)
}
\arguments{
\item{ir_output}{}
}
\description{
Add missing columns to make inconsistent IR output work
}
