% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleleColumnCheck.R
\name{allelecolumnCheck}
\alias{allelecolumnCheck}
\title{Allele column check}
\usage{
allelecolumnCheck(variant_table)
}
\arguments{
\item{variant_table}{}
}
\value{
Returns boolean or NA based on allele column name.
}
\description{
checks name column to potentially multiply by or not
}
