% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aminoacidConversion.R
\name{amino_acid_conversion_one_to_three}
\alias{amino_acid_conversion_one_to_three}
\title{###### Change one letter to three letter code}
\usage{
amino_acid_conversion_one_to_three(amino_acid_change_entry)
}
\arguments{
\item{amino_acid_change_entry}{}
}
\description{
###### Change one letter to three letter code
}
