% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aminoacidConversion.R
\name{amino_acid_conversion_three_to_one}
\alias{amino_acid_conversion_three_to_one}
\title{###### Change three to one letter code}
\usage{
amino_acid_conversion_three_to_one(amino_acid_change_entry)
}
\arguments{
\item{amino_acid_change_entry}{}
}
\value{

}
\description{
###### Change three to one letter code
}
