% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathowinAnnotation_WatchdogTables.R
\name{annotateWatchdogTables}
\alias{annotateWatchdogTables}
\title{### Annotate files in watchdog folder}
\usage{
annotateWatchdogTables(filepath)
}
\arguments{
\item{filepath}{}
}
\value{

}
\description{
### Annotate files in watchdog folder
}
