% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTSGpos.R
\name{checkTSG}
\alias{checkTSG}
\title{Checks if gene is part of TSG and checks position}
\usage{
checkTSG(gene, aa_pos, TSG_list)
}
\arguments{
\item{TSG_list}{}
}
\value{
'likely pathogenic' if criteria are fullfilled, NA otherwise
}
\description{
Checks if gene is part of TSG and checks position
}
