% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_De_Fr.R
\name{ClassficationDe}
\alias{ClassficationDe}
\title{Translate clinvar clinical significance to german}
\usage{
ClassficationDe(input)
}
\arguments{
\item{input}{}
}
\description{
Translate clinvar clinical significance to german
}
