% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_De_Fr.R
\name{ClassficationFr}
\alias{ClassficationFr}
\title{Translate clinvar clinical significance to french}
\usage{
ClassficationFr(input)
}
\arguments{
\item{input}{}
}
\description{
Translate clinvar clinical significance to french
}
