% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinvar_variant_info_check.R
\name{clinvarCheck}
\alias{clinvarCheck}
\title{Check, download and update Clinvar Summary Table (from NIH)
Checks for presence of Clinvar variant summary file (and it being up-to-date).
If not present (or up-to date, downloads from NIH at absolute path)
Caveat: Absolute paths are specified}
\usage{
clinvarCheck()
}
\arguments{
\item{CLINVAR_SUMMARY_FILEPATH}{Absolute filepath of clinvar summary file}
}
\value{
Boolean indicating if downloaded file has same MD5checksum as MD5checksum on NIH website
}
\description{
Check, download and update Clinvar Summary Table (from NIH)
Checks for presence of Clinvar variant summary file (and it being up-to-date).
If not present (or up-to date, downloads from NIH at absolute path)
Caveat: Absolute paths are specified
}
