% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinvar_variant_info_check.R
\name{clinvarTableOutput}
\alias{clinvarTableOutput}
\title{Combine single edited variants into table}
\usage{
clinvarTableOutput(snv_tb)
}
\arguments{
\item{snv_tb}{}
}
\value{

}
\description{
Combine single edited variants into table
}
