% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvParse.R
\name{cnvParse}
\alias{cnvParse}
\title{Separates CNV confidence column in non-precision panels}
\usage{
cnvParse(cnv_table)
}
\arguments{
\item{cnv_table}{}
}
\value{
Mutation table CNV confidence column edited
}
\description{
Tidy::separates cnv confidence column into two columns.
Bottom 5% and top95% conf. int. limits.
}
