% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aminoacidConversion.R
\name{complex_one_to_three}
\alias{complex_one_to_three}
\title{#### Complex One To Three AA switch}
\usage{
complex_one_to_three(parsestring)
}
\arguments{
\item{parsestring}{}
}
\description{
#### Complex One To Three AA switch
}
