% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_cosmic_tissue.R
\name{cosmic_counter_per_table_element}
\alias{cosmic_counter_per_table_element}
\title{Applies counter_cosmic_variant_tissue rowwise to SNV table}
\usage{
cosmic_counter_per_table_element(table_element)
}
\arguments{
\item{table_element}{Row of SNV table}
}
\value{
List of counter_cosmic_variant_tissue results (COSMIC_n_total, COSMIC_n_tissue)
}
\description{
Applies counter_cosmic_variant_tissue rowwise to SNV table
}
