% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_cosmic_tissue.R
\name{cosmic_counter_wrapper}
\alias{cosmic_counter_wrapper}
\title{Wrapper function for COSMIC tissue count}
\usage{
cosmic_counter_wrapper(snv_table)
}
\arguments{
\item{snv_table}{Unedited watchdog prep_snv table}
}
\value{
snv_table with two additional columns: COSMIC_n_total and COSMIC_n_tissue
}
\description{
Wrapper function for COSMIC tissue count
}
\examples{
prep_snv.txt
}
