% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_cosmic_tissue.R
\name{cosmic_sql_search}
\alias{cosmic_sql_search}
\title{Count variant occurences per tissue in COSMIC db}
\usage{
cosmic_sql_search(gene, change, sql_con_tbl)
}
\arguments{
\item{gene}{Character string containing gene symbol}

\item{change}{Character string with coding or amino acid change}

\item{sql_con_tbl}{Established SQL connection to COSMIC db}
}
\value{
Returns list with total count and tissue specific counts
}
\description{
Count variant occurences per tissue in COSMIC db
}
\examples{
GATA3 p.M294K returns list with
COSMIC_n_tissue breast (43)
COSMIC_n_total 43
}
