% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_mv_GNXS_info.R
\name{cp_mv_gnxs}
\alias{cp_mv_gnxs}
\title{COPY and rename GNXSinfo to GNXS_metadata}
\usage{
cp_mv_gnxs(infofilepath, directory_name)
}
\arguments{
\item{directory_name}{}
}
\description{
COPY and rename GNXSinfo to GNXS_metadata
}
