% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output_files.R
\name{create_output_files}
\alias{create_output_files}
\title{#### Function to write out tables create_output_files
Generates various output files from IonReporter and Genexus files, respectively}
\usage{
create_output_files(list_of_tables, filepath)
}
\arguments{
\item{list_of_tables}{}

\item{filepath}{}
}
\description{
#### Function to write out tables create_output_files
Generates various output files from IonReporter and Genexus files, respectively
}
