% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_snv_AA_position.R
\name{extract_snv_position}
\alias{extract_snv_position}
\title{Extract position of SNVs}
\usage{
extract_snv_position(aa_change)
}
\arguments{
\item{clinvar_ready_AA}{}
}
\value{
SNV amino-acid position
}
\description{
Extract position of SNVs
}
