% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameshift_clinvarfix.R
\name{fsClinvarfix}
\alias{fsClinvarfix}
\title{Frameshift clinvar fix}
\usage{
fsClinvarfix(amino_acid_change)
}
\arguments{
\item{amino_acid_change}{}
}
\description{
Frameshift clinvar fix
}
\examples{
removes 'stop after how many amino-acids' information, e.g.
# p.Ser1465ArgfsTer3 to p.Ser1465fs
}
