% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrecisionRNA_fusionParser.R
\name{FusionOutput}
\alias{FusionOutput}
\title{Generate Combined output file for Fusion.tsv}
\usage{
FusionOutput(fusionpath)
}
\arguments{
\item{fusionpath}{}
}
\value{

}
\description{
Generate Combined output file for Fusion.tsv
}
