% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrecisionRNA_fusionParser.R
\name{fusionParser}
\alias{fusionParser}
\title{Parses Precision RNA Fusion.tsv}
\usage{
fusionParser(fusionpath)
}
\arguments{
\item{filepath}{}
}
\description{
Parses Precision RNA Fusion.tsv
}
