% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_gene_name.R
\name{gene_symbol_sub}
\alias{gene_symbol_sub}
\title{Substitute comma separated gene symbols with single gene name}
\usage{
gene_symbol_sub(gene_symbol)
}
\arguments{
\item{gene_symbol}{}
}
\description{
Substitute comma separated gene symbols with single gene name
}
\examples{
LOC101927447,SCUBE1 becomes SCUBE1
}
