% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_gene_name.R
\name{gene_symbol_sub_table}
\alias{gene_symbol_sub_table}
\title{Substitute gene symbols}
\usage{
gene_symbol_sub_table(snvtable)
}
\arguments{
\item{snvtable}{}
}
\description{
Substitute gene symbols
}
