% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerHotspots_counter.R
\name{info_cancerHotspots}
\alias{info_cancerHotspots}
\title{Retrieve if variant codon position is covered in cancerHotspots (v2)}
\usage{
info_cancerHotspots(geneName, snv_position)
}
\arguments{
\item{snv_position}{}
}
\description{
Retrieve if variant codon position is covered in cancerHotspots (v2)
}
\examples{
Retrieves No. of total entries in CANCER_HOTSPOTS
KRAS p.G12C retrieves n_total == 2175
}
