% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameGNXSoutput.R
\name{infoName}
\alias{infoName}
\title{Extract Sample name information from Genexus/Precision Info.csv}
\usage{
infoName(Infofilepath)
}
\arguments{
\item{filepath}{}
}
\description{
Extract Sample name information from Genexus/Precision Info.csv
}
