% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_SNV_CNV_table.R
\name{make_output_tables}
\alias{make_output_tables}
\title{Generate tables to export}
\usage{
make_output_tables(ir_output)
}
\arguments{
\item{ir_output}{}
}
\value{

}
\description{
Generate tables to export
}
