% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOutputTablesPrecision.R
\name{make_output_tables_precision}
\alias{make_output_tables_precision}
\title{Generates output tables for precision}
\usage{
make_output_tables_precision(snv_indel, cnv)
}
\arguments{
\item{snv_indel}{table containing indels and snvs}

\item{cnv}{table containing cnvs}
}
\description{
Generates output tables for precision
}
