% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aminoacidConversion.R
\name{one_three_aa_precision}
\alias{one_three_aa_precision}
\title{####### Simple One To Three AA amino acid switch, used in aa_parser_singleletter}
\usage{
one_three_aa_precision(trimmed_string)
}
\arguments{
\item{trimmed_string}{}
}
\description{
####### Simple One To Three AA amino acid switch, used in aa_parser_singleletter
}
