% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aminoacidConversion.R
\name{oneORthree_code}
\alias{oneORthree_code}
\title{Single or three letter code}
\usage{
oneORthree_code(amino_acid_change)
}
\arguments{
\item{amino_acid_change}{}
}
\description{
Single or three letter code
}
