% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_ClinicalSignificance.R
\name{orderSignificance}
\alias{orderSignificance}
\title{Order by clinical significance}
\usage{
orderSignificance(diagnosestring)
}
\arguments{
\item{diagnosestring}{}
}
\description{
Order by clinical significance
}
