% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentParse.R
\name{percentParse}
\alias{percentParse}
\title{Percent Frequency parsing}
\usage{
percentParse(percent_frequency)
}
\arguments{
\item{percent_frequency}{}
}
\value{

}
\description{
Percent Frequency parsing
}
\examples{
Converts A=0.08,T=0.00 to 0.08
}
