% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horak_scoring.R
\name{prep_chr}
\alias{prep_chr}
\title{Convert coding and location in prep_snv.txt to gnomad conforming location}
\usage{
prep_chr(prep_snv)
}
\arguments{
\item{prep_snv_path}{}
}
\description{
Convert coding and location in prep_snv.txt to gnomad conforming location
}
