% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepFilepaths.R
\name{prep_filepaths}
\alias{prep_filepaths}
\title{From filepath to watchdog/prep filepaths}
\usage{
prep_filepaths(filepath)
}
\arguments{
\item{filepath}{}
}
\description{
From filepath to watchdog/prep filepaths
}
\examples{
list(prep_snv_filepath, prep_cnv_filepath)
}
