% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadRenameSelect.R
\name{read_rename_select}
\alias{read_rename_select}
\title{read in, rename and select}
\usage{
read_rename_select(filepath)
}
\arguments{
\item{filepath}{}
}
\value{

}
\description{
read in, rename and select
}
