% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadRenameSelect.R
\name{read_rename_select_precision_snv}
\alias{read_rename_select_precision_snv}
\title{Read rename select for GNXS files}
\usage{
read_rename_select_precision_snv(filepath)
}
\arguments{
\item{filepath}{}
}
\description{
Read rename select for GNXS files
}
