% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_list_string.R
\name{row_activityList_format_cnv}
\alias{row_activityList_format_cnv}
\title{Output concatenated string of gene, copy number and chromosome}
\usage{
row_activityList_format_cnv(cnv_prep_row)
}
\arguments{
\item{cnv_prep}{}
}
\value{
Returns one single string of all detected CNV copy numbers
}
\description{
Output concatenated string of gene, copy number and chromosome
}
\examples{
BRAF: 10x (chr7)
}
