% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_list_string.R
\name{row_activityList_format_snv}
\alias{row_activityList_format_snv}
\title{Output concatenated string of gene, change and allele frequency}
\usage{
row_activityList_format_snv(snv_prep_row)
}
\arguments{
\item{snv_prep}{}
}
\value{
Returns one single string
}
\description{
Output concatenated string of gene, change and allele frequency
}
\examples{
ALK: p.D1529E (56.1\%)
}
