% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortenFileName.R
\name{shortenFilename}
\alias{shortenFilename}
\title{Shorten filename}
\usage{
shortenFilename(filepath)
}
\arguments{
\item{filepath}{}
}
\value{
shortened filename
}
\description{
Shorten filename
}
\examples{
W601.B022_v1_44534543535435345.tsv becomes 601.B022
}
