% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_list_string.R
\name{snv_activityList_format}
\alias{snv_activityList_format}
\title{Output single string of variants}
\usage{
snv_activityList_format(snv_prep_filepath)
}
\arguments{
\item{snv_prep}{}
}
\value{

}
\description{
Output single string of variants
}
\examples{
ALK: p.D1529E (56.1\%); EGFR: c.1498+22A>T (47.8\%);
}
