% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snvParse.R
\name{snvParse}
\alias{snvParse}
\title{SNV PARSE
Takes SNV part of IR output and converts entry with comma to long format.
This procedure is necessary to only parse one entry at a time for all downstream processes}
\usage{
snvParse(snv)
}
\arguments{
\item{snv}{}
}
\value{
Edited SNV
}
\description{
SNV PARSE
Takes SNV part of IR output and converts entry with comma to long format.
This procedure is necessary to only parse one entry at a time for all downstream processes
}
