% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliceSiteAnnot.R
\name{splicesite_Annot}
\alias{splicesite_Annot}
\title{### splice site annotation Function to interrogate if variant occurs in splice site}
\usage{
splicesite_Annot(variant_table)
}
\arguments{
\item{variant_table}{}
}
\description{
### splice site annotation Function to interrogate if variant occurs in splice site
}
