% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entryTable_metadata.R
\name{std_metadataCollector}
\alias{std_metadataCollector}
\title{Collects Metadata and writes out table
This function extracts metadata stored in '##' commented lines from ThermoFisher IonReporter .tsv reports.
It's made to only work with '##' commented files/lines.}
\usage{
std_metadataCollector(filepath)
}
\arguments{
\item{filepath}{Absolute or relative path to IonReporter .tsv file}
}
\value{
A table with one column, namely ## IR Workflow Metainformation, and '##' entries as rows
}
\description{
Collects Metadata and writes out table
This function extracts metadata stored in '##' commented lines from ThermoFisher IonReporter .tsv reports.
It's made to only work with '##' commented files/lines.
}
