% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WatchdogTables_Pipelines.R
\name{stdPipeline_TableOutput}
\alias{stdPipeline_TableOutput}
\title{Pipeline from intro file to Watchdog output files.
These files can be used for further downstream/alternative website processing.}
\usage{
stdPipeline_TableOutput(snv_filepath)
}
\arguments{
\item{snv_filepath}{}
}
\value{

}
\description{
Pipeline from intro file to Watchdog output files.
These files can be used for further downstream/alternative website processing.
}
