% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTSGpos.R
\name{tsgParseTable}
\alias{tsgParseTable}
\title{tsgParseTable Parse SNV which went through 'watchdog annotation' step}
\usage{
tsgParseTable(snvtable)
}
\arguments{
\item{snvtable}{}
}
\value{
snvtable with tsgInfo column added
}
\description{
tsgParseTable Parse SNV which went through 'watchdog annotation' step
}
