% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerHotspots_counter.R
\name{wrapper_row_cancerHotspots}
\alias{wrapper_row_cancerHotspots}
\title{Apply cancerhotspot check to single row}
\usage{
wrapper_row_cancerHotspots(variantRow)
}
\arguments{
\item{variantRow}{}
}
\description{
Apply cancerhotspot check to single row
}
