% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerHotspots_counter.R
\name{wrapper_table_cancerHotspots}
\alias{wrapper_table_cancerHotspots}
\title{Apply cancerhotspot check to snv_variant table}
\usage{
wrapper_table_cancerHotspots(variantTable)
}
\arguments{
\item{variantTable}{}
}
\description{
Apply cancerhotspot check to snv_variant table
}
