% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTSGpos.R
\name{checkTSG}
\alias{checkTSG}
\title{Check if Gene is TSG and if variant leads to Ter or frameshift within 90% of AA length}
\usage{
checkTSG(gene, aa_pos, TSG_list)
}
\arguments{
\item{TSG_list}{}
}
\value{
'likely pathogenic' if criteria are fullfilled, NA otherwise
}
\description{
Check if Gene is TSG and if variant leads to Ter or frameshift within 90% of AA length
}
