% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_variants_per_panel.R
\name{count_variants_per_panel}
\alias{count_variants_per_panel}
\title{# COUNT Variants per panel -- only SNVs for now.}
\usage{
count_variants_per_panel(prep_tbl, SNV_COUNT)
}
\arguments{
\item{SNV_COUNT}{}
}
\description{
# COUNT Variants per panel -- only SNVs for now.
}
