% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_cosmic_tissue.R
\name{counter_cosmic_variant_tissue}
\alias{counter_cosmic_variant_tissue}
\title{Count variant occurences per tissue in COSMIC db#'}
\usage{
counter_cosmic_variant_tissue(variants_per_tissue)
}
\arguments{
\item{variants_per_tissue}{}
}
\description{
Count variant occurences per tissue in COSMIC db#'
}
