% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_SNV_CNV_table.R
\name{make_output_tables}
\alias{make_output_tables}
\title{## IDEA split IRoutput by type and annotate respectively}
\usage{
make_output_tables(ir_output)
}
\arguments{
\item{ir_output}{}
}
\value{

}
\description{
## IDEA split IRoutput by type and annotate respectively
}
