% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTSGpos.R
\name{tsgParseTable}
\alias{tsgParseTable}
\title{tsgParseTable Checks SNV contains 'fs' or 'Ter'}
\usage{
tsgParseTable(snvtable)
}
\arguments{
\item{snvtable}{}
}
\value{
snvtable with tsgInfo column added
}
\description{
tsgParseTable Checks SNV contains 'fs' or 'Ter'
}
